CatalogueRescan: if opendir() fails, unref fpes, but don't free the cat
None of the callers of CatalogueRescan check for failure before accessing
the cat pointer so don't free it (especially without clearing the pointer
to it in fpe->private), just unref the contents.

Can only be triggered if somehow stat() succeeds on the directory, but
opendir fails anyway (removed between the calls?  permission problem?).

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>